Replace mongodb with home-grown database
Since this project is supposed to run on a Raspberry Pi v1 it
cannot use Mongodb (Mongodb is 64bit only). Introduce a new db
that recides in memory and is async only on the disk io. Will not
go well with a lot of data but is good enough for now.